+++ NEWS.md
@@ -1,3 +1,10 @@
+# MsBackendMetabolomicsWorkbench 0.99
+
+## Changes in 0.99.0
+
+- Remove `progress` import.
+- Update to 0.99.0 for Bioconductor submission.
+
 # MsBackendMetabolomicsWorkbench 0.1
 
 ## Changes in 0.1.8